Cleanup the repl function
We had added a few prints to test the repl. These are being removed.

NOTE: there is still a persisting bug which is that the check
for (quit) (exit) etc. doesnt work in general. I have no explaination
for this as of now but this has to be dealth with in the future